Move TryStartKubelet outside preflight checks
